+++ examples/stateless-async/stateless-async-consumer/src/main/java/io/joynr/examples/statelessasync/Bootstrap.java
@@ -42,7 +42,6 @@ public class Bootstrap {
         Bootstrap.logger.info("Starting consumer ...");
         Properties joynrProperties = new Properties();
         joynrProperties.put("joynr.messaging.mqtt.brokeruris", "tcp://mqttbroker:1883");
-        joynrProperties.put(MessagingPropertyKeys.PROPERTY_MESSAGING_PRIMARYGLOBALTRANSPORT, "mqtt");
         joynrProperties.setProperty(MessagingPropertyKeys.PERSISTENCE_FILE, "consumer-joynr.properties");
         joynrProperties.setProperty(PROPERTY_JOYNR_DOMAIN_LOCAL, "stateless_async_consumer_local_domain");
         Module runtimeModule = Modules.combine(new CCInProcessRuntimeModule(), new HivemqMqttClientModule());