+++ java/integration-tests/src/test/java/io/joynr/integration/RoutingTableOverwriteEnd2EndTest.java
@@ -83,7 +83,6 @@ public class RoutingTableOverwriteEnd2EndTest {
 
         mqttConfig = new Properties();
         mqttConfig.put(MqttModule.PROPERTY_MQTT_BROKER_URIS, "tcp://localhost:" + mqttBrokerPort);
-        mqttConfig.put(MessagingPropertyKeys.PROPERTY_MESSAGING_PRIMARYGLOBALTRANSPORT, "mqtt");
 
         providerDomain = "testDomain" + System.currentTimeMillis();
     }