+++ java/messaging/messaging-common/src/test/java/io/joynr/messaging/routing/AddressManagerTest.java
@@ -21,7 +21,7 @@ package io.joynr.messaging.routing;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
-import static org.mockito.Mockito.any;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -48,7 +48,6 @@ import joynr.system.RoutingTypes.MqttAddress;
 
 @RunWith(MockitoJUnitRunner.class)
 public class AddressManagerTest {
-    private static final String NO_PRIMARY_GLOBAL_TRANSPORT_SELECTED = null;
     private static final String GLOBAL_TRANSPORT_MQTT = "mqtt";
 
     @Mock
@@ -83,7 +82,7 @@ public class AddressManagerTest {
 
     @Test
     public void testNoAddressAvailableForMulticast() {
-        createAddressManager(NO_PRIMARY_GLOBAL_TRANSPORT_SELECTED);
+        createAddressManager();
 
         Map<Address, Set<String>> result = subject.getParticipantIdMap(joynrMessage);
         assertEquals(0, result.size());
@@ -91,7 +90,7 @@ public class AddressManagerTest {
 
     @Test
     public void testGetMulticastAddressFromSingleCalculatorForGloballyVisibleProvider() {
-        createAddressManager(NO_PRIMARY_GLOBAL_TRANSPORT_SELECTED, multicastAddressCalculator);
+        createAddressManager(multicastAddressCalculator);
 
         when(routingTable.getIsGloballyVisible(participantId)).thenReturn(true);
         when(joynrMessage.getRecipient()).thenReturn(participantId + "/multicastname");
@@ -111,36 +110,14 @@ public class AddressManagerTest {
     }
 
     @Test(expected = JoynrIllegalStateException.class)
-    public void testMultipleCalculatorsNoPrimaryGlobalTransport() {
+    public void testMultipleCalculators() {
         MulticastAddressCalculator anotherMulticastAddressCalculator = mock(MulticastAddressCalculator.class);
-        createAddressManager(NO_PRIMARY_GLOBAL_TRANSPORT_SELECTED,
-                             multicastAddressCalculator,
-                             anotherMulticastAddressCalculator);
-    }
-
-    @Test
-    public void testGetAddressFromMultipleCalculators() {
-        MulticastAddressCalculator anotherMulticastAddressCalculator = mock(MulticastAddressCalculator.class);
-        when(anotherMulticastAddressCalculator.supports(GLOBAL_TRANSPORT_MQTT)).thenReturn(true);
-        when(anotherMulticastAddressCalculator.calculate(joynrMessage)).thenReturn(multicastAddresses);
-        createAddressManager(GLOBAL_TRANSPORT_MQTT, multicastAddressCalculator, anotherMulticastAddressCalculator);
-
-        when(joynrMessage.getRecipient()).thenReturn(participantId + "/multicastname");
-
-        Map<Address, Set<String>> result = subject.getParticipantIdMap(joynrMessage);
-        assertNotNull(result);
-        assertEquals(1, result.size());
-        assertEquals(1, result.values().iterator().next().size());
-
-        DelayableImmutableMessage delayablemessage = new DelayableImmutableMessage(joynrMessage,
-                                                                                   1000,
-                                                                                   result.values().iterator().next());
-        assertEquals(Optional.of(multicastAddress), subject.getAddressForDelayableImmutableMessage(delayablemessage));
+        createAddressManager(multicastAddressCalculator, anotherMulticastAddressCalculator);
     }
 
     @Test
     public void testGetLocalMulticastParticipantAddresses() {
-        createAddressManager(GLOBAL_TRANSPORT_MQTT, multicastAddressCalculator);
+        createAddressManager(multicastAddressCalculator);
 
         when(joynrMessage.isReceivedFromGlobal()).thenReturn(true);
         String multicastId = "from/to";
@@ -180,7 +157,7 @@ public class AddressManagerTest {
 
     @Test
     public void testGetLocalMulticastParticipantWithoutGlobalTransports() {
-        createAddressManager(NO_PRIMARY_GLOBAL_TRANSPORT_SELECTED);
+        createAddressManager();
 
         String multicastId = "from/to";
         when(joynrMessage.getRecipient()).thenReturn(multicastId);
@@ -202,7 +179,7 @@ public class AddressManagerTest {
 
     @Test
     public void testGetLocalAndGlobalAddressesForGloballyVisibleProvider() {
-        createAddressManager(GLOBAL_TRANSPORT_MQTT, multicastAddressCalculator);
+        createAddressManager(multicastAddressCalculator);
 
         String multicastId = participantId + "/to";
         when(joynrMessage.getSender()).thenReturn(participantId);
@@ -241,7 +218,7 @@ public class AddressManagerTest {
 
     @Test
     public void testGetLocalAndGlobalAddressesForGloballyInvisibleProvider() {
-        createAddressManager(GLOBAL_TRANSPORT_MQTT, multicastAddressCalculator);
+        createAddressManager(multicastAddressCalculator);
 
         when(joynrMessage.getSender()).thenReturn("participantId");
         String multicastId = participantId + "/to";
@@ -270,7 +247,7 @@ public class AddressManagerTest {
 
     @Test
     public void testGetLocalAndGlobalAddressesForNonExistingProvider() {
-        createAddressManager(GLOBAL_TRANSPORT_MQTT, multicastAddressCalculator);
+        createAddressManager(multicastAddressCalculator);
 
         when(joynrMessage.getSender()).thenReturn("participantId");
         String multicastId = participantId + "/to";
@@ -299,7 +276,7 @@ public class AddressManagerTest {
 
     @Test
     public void testGetMulticastAddressFromSingleCalculatorForGloballyInvisibleProvider() {
-        createAddressManager(NO_PRIMARY_GLOBAL_TRANSPORT_SELECTED, multicastAddressCalculator);
+        createAddressManager(multicastAddressCalculator);
 
         when(joynrMessage.getSender()).thenReturn(participantId + "/multicastname");
         when(multicastAddressCalculator.createsGlobalTransportAddresses()).thenReturn(true);
@@ -310,7 +287,7 @@ public class AddressManagerTest {
 
     @Test
     public void testGetMulticastAddressFromSingleCalculatorForGloballyInvisibleProviderNoGlobalTransport() {
-        createAddressManager(NO_PRIMARY_GLOBAL_TRANSPORT_SELECTED, multicastAddressCalculator);
+        createAddressManager(multicastAddressCalculator);
 
         when(multicastAddressCalculator.createsGlobalTransportAddresses()).thenReturn(false);
         when(multicastAddressCalculator.calculate(joynrMessage)).thenReturn(multicastAddresses);
@@ -322,7 +299,7 @@ public class AddressManagerTest {
 
     @Test
     public void testGetSeveralMulticastAddressesFromSingleCalculatorForGloballyVisibleProvider() {
-        createAddressManager(NO_PRIMARY_GLOBAL_TRANSPORT_SELECTED, multicastAddressCalculator);
+        createAddressManager(multicastAddressCalculator);
 
         MqttAddress secondMulticastAddress = new MqttAddress("testgbid", "testtopic");
         multicastAddresses.add(secondMulticastAddress);
@@ -354,7 +331,7 @@ public class AddressManagerTest {
 
     @Test
     public void testGetMulticastAddressGlobalVisibilityCheckThrowsException() {
-        createAddressManager(NO_PRIMARY_GLOBAL_TRANSPORT_SELECTED, multicastAddressCalculator);
+        createAddressManager(multicastAddressCalculator);
 
         when(joynrMessage.getSender()).thenReturn(participantId + "/multicastname");
         when(multicastAddressCalculator.createsGlobalTransportAddresses()).thenReturn(true);
@@ -363,10 +340,8 @@ public class AddressManagerTest {
         assertEquals(0, result.size());
     }
 
-    private void createAddressManager(String primaryGlobalTransport,
-                                      MulticastAddressCalculator... multicastAddressCalculators) {
+    private void createAddressManager(MulticastAddressCalculator... multicastAddressCalculators) {
         subject = new AddressManager(routingTable,
-                                     new AddressManager.PrimaryGlobalTransportHolder(primaryGlobalTransport),
                                      new HashSet<MulticastAddressCalculator>(Arrays.asList(multicastAddressCalculators)),
                                      multicastReceiverRegistry);
     }