+++ tests/graceful-shutdown-test/graceful-shutdown-test-consumer/src/main/java/io/joynr/tests/gracefulshutdown/Bootstrap.java
@@ -43,7 +43,6 @@ public class Bootstrap {
         logger.info("Starting consumer ...");
         Properties joynrProperties = new Properties();
         joynrProperties.put(PROPERTY_MQTT_BROKER_URIS, "tcp://mqttbroker:1883");
-        joynrProperties.put(MessagingPropertyKeys.PROPERTY_MESSAGING_PRIMARYGLOBALTRANSPORT, "mqtt");
         joynrProperties.setProperty(MessagingPropertyKeys.PERSISTENCE_FILE, "consumer-joynr.properties");
         joynrProperties.setProperty(PROPERTY_JOYNR_DOMAIN_LOCAL, "gracefulshutdown_consumer_local_domain");
         Module runtimeModule = Modules.combine(new CCInProcessRuntimeModule(), new HivemqMqttClientModule());